% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr_equilibrium_plot.R
\name{hcr_equilibrium_plot}
\alias{hcr_equilibrium_plot}
\title{XXX}
\usage{
hcr_equilibrium_plot(df)
}
\arguments{
\item{df}{data.frame, normally generated via function hcr_summarise_data}
}
\description{
XXX
}
