% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr_recruitment_model.R
\name{hcr_recruitment_model}
\alias{hcr_recruitment_model}
\title{HCR recruitment model}
\usage{
hcr_recruitment_model(ssb, cv, ctr)
}
\arguments{
\item{ssb}{XXX}

\item{cv}{XXX}

\item{ctr}{XXX}
}
\description{
XXX
}
