% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr_set_recErrors.R
\name{hcr_set_recErrors}
\alias{hcr_set_recErrors}
\title{HCR: Setup of recruitment error structure}
\usage{
hcr_set_recErrors(d, ctr)
}
\arguments{
\item{d}{XXX}

\item{ctr}{XXX}
}
\description{
XXX
}
