% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit2.R
\name{recruit2}
\alias{recruit2}
\title{Bootstrap model}
\usage{
recruit2(ssb, ssbcut, recmean, rdev)
}
\arguments{
\item{ssb}{XXX}

\item{ssbcut}{XXX}

\item{recmean}{XXX}

\item{rdev}{XXX}
}
\description{
Just a dummy for now
}
