% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{hcr_recruitment_model}
\alias{hcr_recruitment_model}
\title{HCR recruitment model}
\usage{
hcr_recruitment_model(ssb, cv, ctr)
}
\arguments{
\item{ssb}{XXX}

\item{cv}{XXX}

\item{ctr}{XXX}
}
\description{
XXX
}

