% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{hcr_set_recErrors2}
\alias{hcr_set_recErrors2}
\title{HCR: Setup of recruitment error structure}
\usage{
hcr_set_recErrors2(d, ctr)
}
\arguments{
\item{d}{XXX}

\item{ctr}{XXX}
}
\description{
Each iter has own cv and rho
}

